U of A Orbits: UTC time corrected
This commit addresses an identified issue where the UTC time reported
in the file was incorrect.
The offset between TAI time recorded in the CryoSat-2 file and UTC time
was not removed but doubled in the previous version of orbit location
files posted here.
The issue has been closed and the TAI to UTC offset has been correctly
removed, thus the time reported in the file is now UTC time. Please see
the now closed issue report for more details.